// graphql/schema/generated/models/book.model.ts
/* Do not edit this file. It was generated programmatically. */

import { DEFAULT_MODEL_OPTIONS } from "@/graphql/schema/constants";
import definition from "@/graphql/schema/definitions/Book";
import * as Scalars from "@/graphql/schema/scalars";
import * as Types from "@/graphql/schema/types";
import { Model } from "@/graphql/schema/types";
import { getModelForClass, ModelOptions, post, pre, prop as Property } from "@typegoose/typegoose";
import * as TypeGraphQL from "type-graphql-v2-fork";

@TypeGraphQL.ObjectType()
@ModelOptions(DEFAULT_MODEL_OPTIONS)
@pre<Book>(
  "save",
  (definition?.hooks?.save?.pre as any) ??
    async function (next) {
      return next();
    }
)
@post<Book>(
  "save",
  definition?.hooks?.save?.post ??
    (async (_instance) => {
      return;
    })
)
export class Book extends Model {
  @TypeGraphQL.Field(() => Scalars.String, { nullable: true })
  @Property({ type: () => String, required: false, default: null })
  isbn?: Types.String | null;

  @TypeGraphQL.Field(() => Scalars.String, { nullable: true })
  @Property({ type: () => String, required: false, default: null })
  isbn13?: Types.String | null;

  @TypeGraphQL.Field(() => Scalars.String, { nullable: false })
  @Property({ type: () => String, required: true })
  title!: Types.String;

  @TypeGraphQL.Field(() => Scalars.String, { nullable: false })
  @Property({ type: () => String, required: true })
  slug!: Types.String;

  @TypeGraphQL.Field(() => Scalars.String, { nullable: true })
  @Property({ type: () => String, required: false, default: null })
  description?: Types.String | null;

  @TypeGraphQL.Field(() => Scalars.StringArray, { nullable: false })
  @Property({ required: true })
  authorNames!: Types.String[];

  @TypeGraphQL.Field(() => Scalars.StringArray, { nullable: false })
  @Property({ required: true })
  authorNamesLf!: Types.String[];

  @TypeGraphQL.Field(() => Scalars.Int, { nullable: true })
  @Property({ required: false, default: null })
  publicationYear?: Types.Number | null;

  @TypeGraphQL.Field(() => Scalars.Int, { nullable: true })
  @Property({ required: false, default: null })
  originalPublicationYear?: Types.Number | null;
}

const BookModel = getModelForClass<typeof Book>(Book);
export default BookModel;